Refactor out District and Area information of Singapore
